organize stacked borrows run-pass tests
