+++ tests/run-pass/stacked-borrows/2phase.rs
