+++ tests/run-pass/stacked-borrows/stacked-borrows.rs
